backend/requirements.txt
fastapi
uvicorn
google-genai
python-dotenv
pydantic
dotenv